add functional test for new avro based reader context
